Test-Path

+++ Deploy-Application.ps1
@@ -127,8 +127,10 @@ Try {
 		Show-InstallationProgress
 
 		## <Perform Pre-Installation tasks here>
-		$exitCode = Execute-Process -Path "$envProgramFiles\ANSYS Inc\v190\Uninstall.exe" -Parameters "-silent" -WindowStyle "Hidden" -PassThru
-        If (($exitCode.ExitCode -ne "0") -and ($mainExitCode -ne "3010")) { $mainExitCode = $exitCode.ExitCode }
+		If (Test-Path -Path $envProgramFiles\ANSYS Inc\v190) {
+			$exitCode = Execute-Process -Path "$envProgramFiles\ANSYS Inc\v190\Uninstall.exe" -Parameters "-silent" -WindowStyle "Hidden" -PassThru
+			If (($exitCode.ExitCode -ne "0") -and ($mainExitCode -ne "3010")) { $mainExitCode = $exitCode.ExitCode }
+		}
 
 		##*===============================================
 		##* INSTALLATION
@@ -188,9 +190,10 @@ Try {
 		Execute-Process -Path "$envProgramFiles\ANSYS Inc\v202\Uninstall.exe" -Parameters "-silent" -WindowStyle "Hidden" -PassThru
 		## $exitCode = Execute-Process -Path "$envProgramFiles\ANSYS Inc\v202\Uninstall.exe" -Parameters "-silent" -WindowStyle "Hidden" -PassThru
 		## If (($exitCode.ExitCode -ne "0") -and ($mainExitCode -ne "3010")) { $mainExitCode = $exitCode.ExitCode }
-		Get-ChildItem -Path "$envProgramFiles\ANSYS Inc\" -Recurse | Remove-Item -Force -Recurse
-		Remove-Item -Path "$envProgramFiles\ANSYS Inc\" -Force
-
+		If (Test-Path -Path $envProgramFiles\ANSYS Inc) {
+			Get-ChildItem -Path "$envProgramFiles\ANSYS Inc\" -Recurse | Remove-Item -Force -Recurse
+			Remove-Item -Path "$envProgramFiles\ANSYS Inc\" -Force
+		}
 		##*===============================================
 		##* POST-UNINSTALLATION
 		##*===============================================
